Adding poller for results,, updated diagram

+++ mongodb/src/main/kotlin/com/example/mongodb/model/TimingsSummary.kt
@@ -7,5 +7,5 @@ data class TimingsSummary (
     val max: Instant?,
     val count: Int,
     val totalTime: Long?,
-    val avgTime: Long?
+    val avgTime: Float?
 )
